requirements.txt
alembic==1.12.1
bcrypt==4.0.1
blinker==1.7.0
click==8.1.7
dnspython==2.4.2
email-validator==2.1.0.post1
flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-CKEditor==0.5.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
flask-sqlalchemy==3.1.1
flask-wtf==1.2.1
greenlet==3.0.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
Pillow==10.1.0
PyJWT==2.8.0
SQLAlchemy==2.0.23
typing-extensions==4.8.0
werkzeug==3.0.1
wtforms==3.1.1
zipp==3.17.0
